+++ src/main/java/ca/ioxom/ioxbot/commands/CommandMethods.java
@@ -1,10 +1,10 @@
-package ca.ioxom.ioxbot.stuff;
+package ca.ioxom.ioxbot.commands;
 
-import ca.ioxom.ioxbot.Config;
-import ca.ioxom.ioxbot.Main;
+import ca.ioxom.ioxbot.frame.Main;
+import ca.ioxom.ioxbot.other.Config;
 import net.dv8tion.jda.api.events.message.MessageReceivedEvent;
 
-public class ExtraCommands {
+public class CommandMethods {
     public static void bullyAlex(MessageReceivedEvent event) {
         if ((event.getAuthor().getId().equals("382659726778957858") || event.getAuthor().getId().equals("730866562189230110")) && (event.getMessage().getContentRaw().contains("https://www.youtu")
                 || event.getMessage().getContentRaw().contains("www.youtube"))) {
